Added player movement.

+++ server/DatabaseBuilder.cs
@@ -32,8 +32,8 @@ internal class DatabaseBuilder
 	            'Strength'	INTEGER NOT NULL DEFAULT 10,
 	            'Speed'	INTEGER NOT NULL DEFAULT 10,
 	            'Map_id'	INTEGER NOT NULL DEFAULT 1,
-	            'X_Coordinate'	INTEGER NOT NULL DEFAULT 0,
-	            'Y_Coordinate'	INTEGER NOT NULL DEFAULT 0,
+	            'X_Coordinate'	REAL NOT NULL DEFAULT 0,
+	            'Y_Coordinate'	REAL NOT NULL DEFAULT 0,
 	            'Direction'	INTEGER NOT NULL DEFAULT 0,
 	            PRIMARY KEY('User_Id' AUTOINCREMENT)
             );";
